requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
dlib==19.24.6
dnspython==2.7.0
email_validator==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
numpy==2.2.1
pillow==11.0.0
python-dotenv==1.0.1
reportlab==4.2.5
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
